Move Pyramidtype to enums module
