+++ src/rok4/enums.py
@@ -4,6 +4,13 @@
 from enum import Enum
 
 
+class PyramidType(Enum):
+    """Pyramid's data type"""
+
+    RASTER = "RASTER"
+    VECTOR = "VECTOR"
+
+
 class StorageType(Enum):
     """Matrice de correspondance entre type de stockage et protocole."""
